refactored input component
