+++ src/features/usePathname.js
@@ -1,0 +1,7 @@
+import { useLocation } from "react-router-dom";
+
+export const usePathname = () => {
+    const location = useLocation();
+
+    return location.pathname;
+};
